Add GrowingGridPositionStrategy
Rename GridPositionStrategy to FixedGridPositionStrategy to
disambiguate.

GrowingGridPositionStrategy was an experiment in making a
GridPositionStrategy that didn't have a fixed size, but it doesn't work
with my sprites. They aren't square and expect to be overlapped, so they
need to be spawned in a specific order. I could probably fix it if I set
the z position, but that seems too complicated to be worth it.