Use CreatureControl for both flocking and avoidance
Refactor CreatureControl so it can be attached to child objects and they
add forces to their parent. This allows us to have separate flocking and
avoidance child objects with different sized Colliders. We can also put
each child object on a different layer so we don't need to check the
type of the object we're colliding with -- it only collides with the
right kind of object.