Add HasNext to PositionStrategy
Since FixedGridPositionStrategy supports a limited number of positions
and it's not crazy to expect other positions to do the same, add
HasNext.

This also fixes my performance problem where I forget to reduce
numToSpawn and the game slows down because many tiles are spawned on top
of each other. I'm surprised there's no exception when you call MoveNext
on an IEnumerator that's at its end.