Add an enclosure that stops avoidance
The enclosure is a pen (like a pig pen not a writing pen).

The enclosure can contain creatures. When they enter, their Obstacle
Finder Flocking component is disabled so they won't try to leave the
enclosure unless they're following other creatures.

Allows Flocking to be disabled. Draws a black cross to show it's been
turned off.

Adds LayerUtil to help deal with layers and bitmasks. (I already screwed
up the comparison.)

Adds ToggleFlockingArea that can turn flocking on and off on entry/exit
of an area. You can specify a target layer so that you can control which
Flocking you toggle.

I don't know why it's using a Rect Transform. It just happened.