Only eat when wolf reaches a boundary
Use the tags to see if we're at a boundary. I'm pulling the tag from the
Boundary object, so maybe I should use some kind of prefab matching, but
this works for now and is more flexible if there are multiple "forest"
type objects (although I'm not sure I'd need more than just the one).