Add expressions and equations
Now multiple pens are connected together by an operator and you can
put sheep into the pens to satisfy the equation.

Placeholder: Changes all text to "Woo Hoo" as soon as equation is true
(even if it's all zeroes).

Next: Require all sheep in pen (and reduce the number of sheep).